TextPreprocessing-main/requirements.txt
PyYAML
pandas
numpy
num2words
google-transliteration-api
